backend/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Deprecated==1.2.13
dj-rest-auth==3.0.0
Django==4.1.7
django-allauth==0.52.0
django-cors-headers==3.14.0
django-phone-verify==3.0.0
django-phonenumber-field==7.0.2
django-rest-auth==0.9.5
django-trench==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
nexmo==2.5.2
oauthlib==3.2.2
packaging==23.0
phonenumbers==8.13.7
Pillow==9.4.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyotp==2.8.0
pyparsing==3.0.9
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
smsapi-client==2.7.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
twilio==7.16.5
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
wrapt==1.15.0
yubico-client==1.13.0
